Extracted the bold descriptions into their own component, to be used in info and taxonomy.
