+++ projects/laji/src/app/+taxonomy/taxon/info-card/shared/component/taxon-bold/taxon-bold.component.scss
@@ -1,0 +1,3 @@
+.bold-entry {
+  margin-bottom: 5px;
+}
