+++ projects/laji/src/app/+taxonomy/taxon/info-card/taxon-overview/taxon-info/taxon-info.component.ts
@@ -1,12 +1,6 @@
-import { ChangeDetectionStrategy, Component, Input, OnInit, OnDestroy, SimpleChanges, OnChanges } from '@angular/core';
+import { ChangeDetectionStrategy, Component, Input, OnInit, OnDestroy } from '@angular/core';
 import { TranslateService } from '@ngx-translate/core';
 import { Taxonomy } from '../../../../../shared/model/Taxonomy';
-
-interface BoldEntry {
-  scientificName: string;
-  publicRecords: number;
-  bins: string[];
-}
 
 @Component({
   selector: 'laji-taxon-info',
@@ -14,14 +8,13 @@ interface BoldEntry {
   styleUrls: ['./taxon-info.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class TaxonInfoComponent implements OnInit, OnDestroy, OnChanges {
+export class TaxonInfoComponent implements OnInit, OnDestroy {
 
   @Input() taxon: Taxonomy;
 
   langs = ['fi', 'sv', 'en', 'se', 'ru'];
   availableVernacularNames = [];
   availableTaxonNames = {vernacularNames: [], colloquialVernacularNames: []};
-  boldEntries: BoldEntry[] = [];
 
   constructor(
     public translate: TranslateService
@@ -29,12 +22,6 @@ export class TaxonInfoComponent implements OnInit, OnDestroy, OnChanges {
 
   ngOnInit() {
       this.initLangTaxonNames();
-  }
-
-  ngOnChanges(changes: SimpleChanges): void {
-    if (changes.taxon) {
-      this.getBoldEntries();
-    }
   }
 
   ngOnDestroy() {
@@ -50,48 +37,5 @@ export class TaxonInfoComponent implements OnInit, OnDestroy, OnChanges {
       this.availableTaxonNames.colloquialVernacularNames.push({lang: value});
     }
    });
-  }
-
-  getBoldEntries() {
-    if (this.taxon.hasBold === false) {
-      this.boldEntries = [];
-      return;
-    }
-
-    this.getBoldEntry(this.taxon);
-
-    if (this.taxon.synonyms) {
-      this.taxon.synonyms.forEach(synonym => {
-        this.getBoldEntry(synonym);
-      });
-    }
-  }
-
-  getBoldEntry(taxon: Taxonomy) {
-    let scientificName: string = this.taxon.scientificName;
-    let publicRecords = 0;
-    const bins: string[] = [];
-
-    if (taxon.scientificName) {
-      scientificName = taxon.scientificName;
-    }
-
-    if (taxon.bold?.publicRecords) {
-      publicRecords = taxon.bold.publicRecords;
-    }
-
-    if (taxon.bold?.binCount > 0) {
-      taxon.bold?.bins.forEach(bin => {
-        bins.push(bin);
-      });
-    }
-
-    if (publicRecords > 0 || bins.length > 0) {
-      this.boldEntries.push({
-        scientificName,
-        publicRecords,
-        bins
-      });
-    }
   }
 }
