+++ projects/laji/src/app/+taxonomy/taxon/info-card/taxon-taxonomy/taxon-taxonomy.component.html
@@ -68,7 +68,7 @@
          [title]="'taxonomy.bold' | translate"
          [open]="true"
          [autoToggle]="true">
-      
+      <laji-taxon-bold [taxon]="taxon"></laji-taxon-bold>
     </div>
     <div class="panel panel-default"
          *ngIf="taxon.taxonExpert"
